# RF: Optimize ICC_rep_anova with a memoized helper function