@bbfrederick 2022-04-26
Sorry it took so long to get back to you - kept finding myself on mobile with no screen big enough to look at code.

Yes, that looks like a good switch - I wasn't aware of lru_cache (I need to get out more)!  Maxsize=1 is fine - my use case is calculating it on ~260k voxels in a set of 3D images from 1000 subjects with 4 conditions - exactly the same matrix size each time, so the repeated pseudoinverse was killing me....  I don't imagine you'd need to switch the matrix up very frequently, although I'm sure you could come up with some pathological application that would.